[examples] change example platform namespacing
