+++ Android.mk
@@ -180,8 +180,8 @@ LOCAL_SRC_FILES                                          := \
     src/posix/platform/diag.c                               \
     src/posix/platform/misc.c                               \
     src/posix/platform/logging.c                            \
-    src/posix/platform/platform.c                           \
     src/posix/platform/random.c                             \
+    src/posix/platform/system.c                             \
     src/posix/platform/uart.c                               \
     src/posix/platform/flash.c                              \
     src/posix/platform/spi-stubs.c                          \
