+++ examples/apps/ncp/main.c
@@ -34,7 +34,7 @@
 #include <openthread/ncp.h>
 #include <openthread/openthread.h>
 
-#include "platform.h"
+#include "openthread-system.h"
 
 #if OPENTHREAD_ENABLE_MULTIPLE_INSTANCES
 void *otPlatCAlloc(size_t aNum, size_t aSize)
@@ -64,7 +64,7 @@ int main(int argc, char *argv[])
 
 pseudo_reset:
 
-    PlatformInit(argc, argv);
+    otSysInit(argc, argv);
 
 #if OPENTHREAD_ENABLE_MULTIPLE_INSTANCES
     // Call to query the buffer size
@@ -87,10 +87,10 @@ pseudo_reset:
     otDiagInit(sInstance);
 #endif
 
-    while (!PlatformPseudoResetWasRequested())
+    while (!otSysPseudoResetWasRequested())
     {
         otTaskletsProcess(sInstance);
-        PlatformProcessDrivers(sInstance);
+        otSysProcessDrivers(sInstance);
     }
 
     otInstanceFinalize(sInstance);
