+++ examples/platforms/da15000/Makefile.am
@@ -63,9 +63,9 @@ PLATFORM_SOURCES                                                                
     flash.c                                                                               \
     logging.c                                                                             \
     misc.c                                                                                \
-    platform.c                                                                            \
     radio.c                                                                               \
     random.c                                                                              \
+    system.c                                                                              \
     uart.c                                                                                \
     crypto/aes_alt.c                                                                      \
     $(NULL)
