+++ examples/platforms/emsk/Makefile.am
@@ -70,9 +70,9 @@ PLATFORM_SOURCES                                                                
     flash.c                                                                                         \
     logging.c                                                                                       \
     misc.c                                                                                          \
-    platform.c                                                                                      \
     radio.c                                                                                         \
     random.c                                                                                        \
+    system.c                                                                                        \
     uart.c                                                                                          \
     platform-emsk.h                                                                                 \
     $(NULL)
