+++ examples/platforms/kw41z/Makefile.am
@@ -54,9 +54,9 @@ PLATFORM_SOURCES                                                                
     flash.c                                                                                          \
     logging.c                                                                                        \
     misc.c                                                                                           \
-    platform.c                                                                                       \
     radio.c                                                                                          \
     random.c                                                                                         \
+    system.c                                                                                         \
     uart.c                                                                                           \
     $(NULL)
