+++ examples/platforms/nrf52840/system.c
@@ -37,8 +37,8 @@
 
 #include <openthread/platform/logging.h>
 
+#include "openthread-system.h"
 #include "platform-nrf5.h"
-#include "platform.h"
 #include <drivers/clock/nrf_drv_clock.h>
 #include <nrf.h>
 
@@ -52,11 +52,11 @@ void __cxa_pure_virtual(void)
         ;
 }
 
-void PlatformInit(int argc, char *argv[])
+void otSysInit(int argc, char *argv[])
 {
     if (gPlatformPseudoResetWasRequested)
     {
-        PlatformDeinit();
+        otSysDeinit();
     }
 
     (void)argc;
@@ -90,7 +90,7 @@ void PlatformInit(int argc, char *argv[])
     gPlatformPseudoResetWasRequested = false;
 }
 
-void PlatformDeinit(void)
+void otSysDeinit(void)
 {
     nrf5TempDeinit();
     nrf5RadioDeinit();
@@ -111,12 +111,12 @@ void PlatformDeinit(void)
 #endif
 }
 
-bool PlatformPseudoResetWasRequested(void)
+bool otSysPseudoResetWasRequested(void)
 {
     return gPlatformPseudoResetWasRequested;
 }
 
-void PlatformProcessDrivers(otInstance *aInstance)
+void otSysProcessDrivers(otInstance *aInstance)
 {
     nrf5AlarmProcess(aInstance);
     nrf5RadioProcess(aInstance);
@@ -127,7 +127,7 @@ void PlatformProcessDrivers(otInstance *aInstance)
 #endif
 }
 
-__WEAK void PlatformEventSignalPending(void)
+__WEAK void otSysEventSignalPending(void)
 {
     // Intentionally empty
 }
