+++ examples/platforms/openthread-system.h
@@ -29,11 +29,11 @@
 /**
  * @file
  * @brief
- *   This file defines the platform-specific initializers.
+ *   This file defines the platform-specific functions needed by OpenThread's example applications.
  */
 
-#ifndef PLATFORM_H_
-#define PLATFORM_H_
+#ifndef OPENTHREAD_SYSTEM_H_
+#define OPENTHREAD_SYSTEM_H_
 
 #include <openthread/types.h>
 
@@ -42,41 +42,59 @@ extern "C" {
 #endif
 
 /**
- * This function performs all platform-specific initialization.
+ * This function performs all platform-specific initialization of OpenThread's drivers.
+ *
+ * @note This function is not called by the OpenThread library. Instead, the system/RTOS should call this function
+ *       when initialization of OpenThread's drivers is most approriate.
+ *
+ * @param[in]  argc  Number of arguments in @p argv.
+ * @param[in]  argv  Argument vector.
  *
  */
-void PlatformInit(int argc, char *argv[]);
+void otSysInit(int argc, char *argv[]);
 
 /**
- * This function performs all platform-specific deinitialization.
+ * This function performs all platform-specific deinitialization for OpenThread's drivers.
+ *
+ * @note This function is not called by the OpenThread library. Instead, the system/RTOS should call this function
+ *       when deinitialization of OpenThread's drivers is most appropriate.
  *
  */
-void PlatformDeinit(void);
+void otSysDeinit(void);
 
 /**
- * This function returns true is a pseudo-reset was requested.
- * In such a case, the main loop should shut down and re-initialize
- * the OpenThread instance.
+ * This function returns true if a pseudo-reset was requested.
+ *
+ * In such a case, the main loop should shut down and re-initialize the OpenThread instance.
+ *
+ * @note This function is not called by the OpenThread library. Instead, the system/RTOS should call this function
+ *       in the main loop to determine when to shut down and re-initialize the OpenThread instance.
  *
  */
-bool PlatformPseudoResetWasRequested(void);
+bool otSysPseudoResetWasRequested(void);
 
 /**
- * This function performs all platform-specific processing.
+ * This function performs all platform-specific processing for OpenThread's example applications.
+ *
+ * @note This function is not called by the OpenThread library. Instead, the system/RTOS should call this function
+ *       in the main loop when processing OpenThread's drivers is most appropriate.
  *
  * @param[in]  aInstance  The OpenThread instance structure.
  *
  */
-void PlatformProcessDrivers(otInstance *aInstance);
+void otSysProcessDrivers(otInstance *aInstance);
 
 /**
  * This function is called whenever platform drivers needs processing.
  *
+ * @note This function is not handled by the OpenThread library. Instead, the system/RTOS should handle this function
+ *       and schedule a call to `otSysProcessDrivers()`.
+ *
  */
-extern void PlatformEventSignalPending(void);
+extern void otSysEventSignalPending(void);
 
 #ifdef __cplusplus
 } // end of extern "C"
 #endif
 
-#endif // PLATFORM_H_
+#endif // OPENTHREAD_SYSTEM_H_
