+++ examples/platforms/posix/misc.c
@@ -35,7 +35,7 @@
 #include <openthread/types.h>
 #include <openthread/platform/misc.h>
 
-#include "platform.h"
+#include "openthread-system.h"
 
 #ifndef _WIN32
 extern int    gArgumentsCount;
@@ -66,7 +66,7 @@ void otPlatReset(otInstance *aInstance)
 
     argv[gArgumentsCount] = NULL;
 
-    PlatformDeinit();
+    otSysDeinit();
     platformUartRestore();
 
     alarm(0);
