+++ examples/platforms/samr21/diag.c
