+++ src/posix/platform/misc.c
@@ -33,7 +33,7 @@
 #include <openthread/types.h>
 #include <openthread/platform/misc.h>
 
-#include "platform.h"
+#include "openthread-system.h"
 
 extern int    gArgumentsCount;
 extern char **gArguments;
@@ -61,7 +61,7 @@ void otPlatReset(otInstance *aInstance)
 
     argv[gArgumentsCount] = NULL;
 
-    PlatformDeinit();
+    otSysDeinit();
     platformUartRestore();
 
     alarm(0);
